Add flowtype ast parsing for rn-codegen
Summary:
Here's the [[ https://phabricator.khanacademy.org/D25646 | first diff ]],
which has the readme, config files, and the main `run.js` file.

Our code generation is based on [[ http://flowtype.org/ | flow ]] type
definitions on the javascript files which define the React Native root
components. The `flow` cli tool has an option to output the parsed AST of a
javascript file in JSON format; we use this to create an internal
representation of the types of things, which we then use to generate the
data-marshalling code.

This diff just contains the flow-ast processing code & the shared type
definitions that we'll be using in various places.

Test Plan: TODO tests coming soon

Reviewers: charlie

Subscribers: benkomalo, #android, #ios

Differential Revision: https://phabricator.khanacademy.org/D25647