Implement code gen for the ViewController's properties
Summary:
The Swift bridging code consists of 2 generated classes; a ViewController
(ish) and a PropsManager. The PropsManager is solely responsible for
serialization/deserialization of the props, and the ViewController initializes
& updates the ReactNative view, and holds onto the state of the current props
(the original values, not the serialized ones).

In this diff, we generate code for the properties that manage the state of the
props. A parent view can then update the props by doing
`someReactNativeView.propName = newValue` and it will be automatically
serialized.

Test Plan: TODO

Reviewers: charlie

Subscribers: andy, #android, #ios

Differential Revision: https://phabricator.khanacademy.org/D25653