Implement swift class generation
Summary:
The Swift bridging code consists of 2 generated classes; a ViewController
(ish) and a PropsManager. The PropsManager is solely responsible for
serialization/deserialization of the props, and the ViewController initializes
& updates the ReactNative view, and holds onto the state of the current props
(the original values, not the serialized ones).

This diff represents the last of the swift codegen code.

Test Plan: TODO (see upcoming diff for example output)

Reviewers: charlie

Subscribers: #android, #ios

Differential Revision: https://phabricator.khanacademy.org/D25657