Add examples + supporting swift files [ReactNative codegen]
Summary:
There are two example react native root components, meant to demonstrate the
full range of react-native-codegen's capabilities, and give an idea for what
the generated code looks like.

There are also some swift classes that are used by the generated code, to
interact with React Native and the JavaScript side.

Test Plan:
Not sure whether it makes sense to have tests for the swift stuff... seems
like it would be really difficult as it depends on React Native stuff.

Reviewers: charlie

Subscribers: bryan, jasonbrennan, nacho, andy, #android, #ios

Differential Revision: https://phabricator.khanacademy.org/D25659